Fix top-level finder requests, and bump version
