+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.7.0-alpha"
+version in ThisBuild := "0.7.0-alpha2"
